move code to events.clients
